improved tanks, added GetNormal() function

+++ SealTeam7/Assets/Scripts/Enemies/FunkyPhysics/BasePhysics.cs
@@ -15,33 +15,42 @@ namespace Enemies.FunkyPhysics
         [SerializeField] protected float fallDeathVelocityY;
         protected Enemy Self;
         protected Rigidbody Rb;
+        protected bool Grounded;
         
         protected virtual void Start()
         {
             Rb = GetComponent<Rigidbody>();
             Self = GetComponent<Enemy>();
-            Rb.freezeRotation = true;
         }
 
-        protected virtual void Update()
+        private void Update()
+        {
+            if (!GameManager.GetInstance().IsGameActive()) return;
+
+            Grounded = transform.position.y < MapManager.GetInstance().GetHeight(transform.position) + groundedOffset;
+            
+            //WOULD DIE BURIED
+            if (transform.position.y < MapManager.GetInstance().GetHeight(transform.position) - sinkFactor) EnemyManager.GetInstance().Kill(Self);
+            //WOULD DIE FALL DMG
+            if (-Rb.linearVelocity.y >= fallDeathVelocityY && Grounded) EnemyManager.GetInstance().Kill(Self);
+
+            EnemyUpdate();
+        }
+
+        private void FixedUpdate()
         {
             if (!GameManager.GetInstance().IsGameActive()) return;
             
-            if (transform.position.y < MapManager.GetInstance().GetHeight(transform.position) - sinkFactor)
-            {
-                //WOULD DIE BURIED
-                EnemyManager.GetInstance().Kill(Self);
-            }
-            else if (Rb.linearVelocity.y > defianceThreshold && transform.position.y < MapManager.GetInstance().GetHeight(transform.position) + groundedOffset)
+            if (Rb.linearVelocity.y > defianceThreshold && Grounded)
             {
                 Physics.Raycast(transform.position, Vector3.down, out var hit, groundedOffset * 2.0f);
                 Rb.AddForce((Vector3.up + hit.normal).normalized * gravityDefiance, ForceMode.Impulse);
             }
-            else if (-Rb.linearVelocity.y >= fallDeathVelocityY && transform.position.y < MapManager.GetInstance().GetHeight(transform.position) + groundedOffset)
-            {
-                //WOULD DIE FALL DMG
-                EnemyManager.GetInstance().Kill(Self);
-            }
+            
+            EnemyFixedUpdate();
         }
+
+        protected virtual void EnemyUpdate() {}
+        protected virtual void EnemyFixedUpdate() {}
     }
 }
